check if file exists before removing it

+++ khan-project1.sh
@@ -1,5 +1,8 @@
 ##!/bin/bash
 
-rm -f khan-project1.db
+if [ -f khan-project1.db ];
+then
+    rm -f khan-project1.db
+fi
 
 cat khan-project1.sql | sqlite3 khan-project1.db
